// src/app/models/Ticket.ts
export class Ticket {
  id: number;
  createTime: string;
  updateTime: string;
  name: string;
  priority: string;
  ticketIssueType: string;
  assignedTo: string;
  expirationDate: string;
  tableListId: number;
  boardId: number;
  createdById: number;
  sprintId: number;
  sequenceNumber: number;
}